+++ packages/terra-clinical-result/CHANGELOG.md
@@ -9,6 +9,7 @@
   * Added screen-reader support for alternative text for when no results are returned.
   * Added screen-reader support for clinical-result FlowsheetResultCell EnteredInError.
   * Added screen-reader support for clinical-result FlowsheetResultCell with multiple results.
+  * Changed FlowsheetResultCell, ResultNameHeaderCell and ResultTimeHeaderCell in clinical-result to use proper semantic html
 
 ## 1.16.0 - (March 29, 2023)